VMS-15920: Fix clean wix distribution build
<==== Mercurial data ====>
Branch: x_VMS-15913_webengine

+++ build_utils/python/qmldeploy.py
@@ -271,7 +271,7 @@ def main():
         if not args.output:
             exit("Output directory is not specified.")
 
-        if args.clean:
+        if args.clean and os.path.exists(args.output):
             shutil.rmtree(args.output)
 
         deploy_util.deploy(
